Adds a grammar file
